Migrate token scope related code to `token_scopes` pkg
