+++ utils/server_auth_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"net/http"
 	"net/http/httptest"
-	"strconv"
 	"testing"
 
 	"github.com/gin-gonic/gin"
@@ -52,44 +51,6 @@ func createContextWithToken(cookieToken, headerToken, queryToken string) *gin.Co
 	ctx.Request = r
 
 	return ctx
-}
-
-func TestScopeContains(t *testing.T) {
-	tests := []struct {
-		name           string
-		tokenScopes    []string
-		expectedScopes []string
-		all            bool
-		want           bool
-	}{
-		{"empty-scopes", []string{}, []string{}, false, false},
-		{"single-match", []string{"read"}, []string{"read"}, false, true},
-		{"no-match", []string{"read"}, []string{"write"}, false, false},
-		{"multiple-matches", []string{"read", "write"}, []string{"read", "write"}, false, true},
-		{"partial-match-all-false", []string{"read", "write"}, []string{"read"}, false, true},
-		{"partial-match-all-true", []string{"read", "write"}, []string{"read"}, true, false},
-		{"case-insensitivity", []string{"Read"}, []string{"read"}, false, true},
-		{"different-lengths-all-true", []string{"read", "write"}, []string{"read"}, true, false},
-		{"exact-match-all-true", []string{"read", "write"}, []string{"write", "read"}, true, true},
-		{"large-input-sets", largeInputSet(), largeInputSet(), false, true},
-		{"nil-inputs", nil, nil, false, false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := scopeContains(tt.tokenScopes, tt.expectedScopes, tt.all); got != tt.want {
-				t.Errorf("scopeContains() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func largeInputSet() []string {
-	var scopes []string
-	for i := 0; i < 1000; i++ {
-		scopes = append(scopes, "scope"+strconv.Itoa(i))
-	}
-	return scopes
 }
 
 func TestCheckAnyAuth(t *testing.T) {
